component Composition solution
