+++ simpleCounter/src/App.jsx
@@ -3,35 +3,25 @@ import { useState } from 'react';
 import Counter from './components/Counter/Counter.jsx';
 import Header from './components/Header.jsx';
 import { log } from './log.js';
+import ConfigureCounter from './components/Counter/ConfigureCounter.jsx';
 
 function App() {
-  log('<App /> rendered');
+	log('<App /> rendered');
+	const [chosenCount, setChosenCount] = useState(0);
 
-  const [enteredNumber, setEnteredNumber] = useState(0);
-  const [chosenCount, setChosenCount] = useState(0);
+	const handleSetCount = (num) => {
+		setChosenCount(num);
+	};
 
-  function handleChange(event) {
-    setEnteredNumber(+event.target.value);
-  }
-
-  function handleSetClick() {
-    setChosenCount(enteredNumber);
-    setEnteredNumber(0);
-  }
-
-  return (
-    <>
-      <Header />
-      <main>
-        <section id="configure-counter">
-          <h2>Set Counter</h2>
-          <input type="number" onChange={handleChange} value={enteredNumber} />
-          <button onClick={handleSetClick}>Set</button>
-        </section>
-        <Counter initialCount={chosenCount} />
-      </main>
-    </>
-  );
+	return (
+		<>
+			<Header />
+			<main>
+				<ConfigureCounter setCount={handleSetCount} />
+				<Counter initialCount={chosenCount} />
+			</main>
+		</>
+	);
 }
 
 export default App;
