+++ simpleCounter/src/components/Counter/ConfigureCounter.jsx
@@ -1,0 +1,24 @@
+import { useState } from 'react';
+
+const ConfigureCounter = ({ setCount }) => {
+	const [enteredNumber, setEnteredNumber] = useState(0);
+
+	function handleChange(event) {
+		setEnteredNumber(+event.target.value);
+	}
+
+	function handleSetClick() {
+		setCount(enteredNumber);
+		setEnteredNumber(0);
+	}
+
+	return (
+		<section id="configure-counter">
+			<h2>Set Counter</h2>
+			<input type="number" onChange={handleChange} value={enteredNumber} />
+			<button onClick={handleSetClick}>Set</button>
+		</section>
+	);
+};
+
+export default ConfigureCounter;
