+++ simpleCounter/src/components/Counter/Counter.jsx
@@ -1,4 +1,4 @@
-import { useState } from 'react';
+import { useState, memo } from 'react';
 
 import IconButton from '../UI/IconButton.jsx';
 import MinusIcon from '../UI/Icons/MinusIcon.jsx';
@@ -7,55 +7,51 @@ import CounterOutput from './CounterOutput.jsx';
 import { log } from '../../log.js';
 
 function isPrime(number) {
-  log(
-    'Calculating if is prime number',
-    2,
-    'other'
-  );
-  if (number <= 1) {
-    return false;
-  }
+	log('Calculating if is prime number', 2, 'other');
+	if (number <= 1) {
+		return false;
+	}
 
-  const limit = Math.sqrt(number);
+	const limit = Math.sqrt(number);
 
-  for (let i = 2; i <= limit; i++) {
-    if (number % i === 0) {
-      return false;
-    }
-  }
+	for (let i = 2; i <= limit; i++) {
+		if (number % i === 0) {
+			return false;
+		}
+	}
 
-  return true;
+	return true;
 }
 
-export default function Counter({ initialCount }) {
-  log('<Counter /> rendered', 1);
-  const initialCountIsPrime = isPrime(initialCount);
+export default memo(function Counter({ initialCount }) {
+	log('<Counter /> rendered', 1);
+	const initialCountIsPrime = isPrime(initialCount);
 
-  const [counter, setCounter] = useState(initialCount);
+	const [counter, setCounter] = useState(initialCount);
 
-  function handleDecrement() {
-    setCounter((prevCounter) => prevCounter - 1);
-  }
+	function handleDecrement() {
+		setCounter((prevCounter) => prevCounter - 1);
+	}
 
-  function handleIncrement() {
-    setCounter((prevCounter) => prevCounter + 1);
-  }
+	function handleIncrement() {
+		setCounter((prevCounter) => prevCounter + 1);
+	}
 
-  return (
-    <section className="counter">
-      <p className="counter-info">
-        The initial counter value was <strong>{initialCount}</strong>. It{' '}
-        <strong>is {initialCountIsPrime ? 'a' : 'not a'}</strong> prime number.
-      </p>
-      <p>
-        <IconButton icon={MinusIcon} onClick={handleDecrement}>
-          Decrement
-        </IconButton>
-        <CounterOutput value={counter} />
-        <IconButton icon={PlusIcon} onClick={handleIncrement}>
-          Increment
-        </IconButton>
-      </p>
-    </section>
-  );
-}
+	return (
+		<section className="counter">
+			<p className="counter-info">
+				The initial counter value was <strong>{initialCount}</strong>. It{' '}
+				<strong>is {initialCountIsPrime ? 'a' : 'not a'}</strong> prime number.
+			</p>
+			<p>
+				<IconButton icon={MinusIcon} onClick={handleDecrement}>
+					Decrement
+				</IconButton>
+				<CounterOutput value={counter} />
+				<IconButton icon={PlusIcon} onClick={handleIncrement}>
+					Increment
+				</IconButton>
+			</p>
+		</section>
+	);
+});
